+++ packages/seeder/src/mock-data.json
@@ -7,8 +7,8 @@
     "carouselProductCount": 6,
     "gridItems": [
       {
-        "header": "Limited Editions",
-        "subheader": "Collector's picks",
+        "header": "Nimara's products",
+        "subheader": "Discover the best vinyl records",
         "headerFontColor": "light",
         "subheaderFontColor": "light",
         "imageUrl": "https://picsum.photos/id/573/600/400"
